Module matrix_sdk_base::deserialized_responses

Structs

AmbiguityChange

A change in ambiguity of room members that an m.room.member event triggers.

AmbiguityChanges

Collection of ambiguioty changes that room member events trigger.

EncryptionInfo

Struct containing information on how an event was decrypted.

JoinedRoom

Updates to joined rooms.

LeftRoom

MembersResponse

A deserialized response for the rooms members API call.

RoomEvent

Rooms

SyncResponse

SyncRoomEvent

A customized version of a room event coming from a sync that holds optional encryption info.

Timeline

Events in the room.

TimelineSlice

A slice of the timeline in the room.

UnreadNotificationsCount

Counts of unread notifications for a room.

Enums

AlgorithmInfo

The algorithm specific information of a decrypted event.

MemberEvent

Wrapper around both MemberEvent-Types

VerificationState

The verification state of the device that sent an event to us.