Trait matrix_sdk_base::events::EventContent

pub trait EventContent: Serialize {
    fn event_type(&self) -> &str;

    fn from_parts(
        event_type: &str, 
        content: Box<RawValue, Global>
    ) -> Result<Self, Error>;
}

The base trait that all event content types implement.

Implementing this trait allows content types to be serialized as well as deserialized.

Required methods

fn event_type(&self) -> &str

A matrix event identifier, like m.room.message.

fn from_parts(
event_type: &str,
content: Box<RawValue, Global>
) -> Result<Self, Error>

Constructs the given event content.

Loading content...