Module matrix_sdk_base::latest_event

Utilities for working with events to decide whether they are suitable for use as a crate::Room::latest_event.

Structs

LatestEvent

Represent all information required to represent a latest event in an efficient way.

Enums

PossibleLatestEvent

Represents a decision about whether an event could be stored as the latest event in a room. Variants starting with Yes indicate that this message could be stored, and provide the inner event information, and those starting with a No indicate that it could not, and give a reason.

Functions

is_suitable_for_latest_event

Decide whether an event could be stored as the latest event in a room. Returns a LatestEvent representing our decision.