Module matrix_sdk_base::locks

Structs

Mutex

An asynchronous Mutex-like type.

MutexGuard

A handle to a held Mutex. The guard can be held across any .await point as it is Send.

RwLock

An asynchronous reader-writer lock.

RwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped.

RwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped.