Module matrix_sdk_base::ruma::api

(De)serializable types for various Matrix APIs requests and responses and abstractions for them.

Modules

client

(De)serializable types for the Matrix Client-Server API. These types can be shared by client and server code.

error

This module contains types for all kinds of errors that can occur when converting between http requests / responses and ruma’s representation of matrix API requests / responses.

Macros

metadata

Convenient constructor for Metadata constants.

Structs

Metadata

Metadata about an API endpoint.

VersionHistory

The complete history of this endpoint as far as Ruma knows, together with all variants on versions stable and unstable.

Enums

AuthScheme

Authentication scheme used by the endpoint.

Direction

The direction to return events from.

MatrixVersion

The Matrix versions Ruma currently understands to exist.

SendAccessToken

An enum to control whether an access token should be added to outgoing requests

VersioningDecision

A versioning “decision” derived from a set of Matrix versions.

Traits

EndpointError

Gives users the ability to define their own serializable / deserializable errors.

IncomingRequest

A request type for a Matrix API endpoint, used for receiving requests.

IncomingResponse

A response type for a Matrix API endpoint, used for receiving responses.

OutgoingRequest

A request type for a Matrix API endpoint, used for sending requests.

OutgoingRequestAppserviceExt

An extension to OutgoingRequest which provides Appservice specific methods.

OutgoingResponse

A request type for a Matrix API endpoint, used for sending responses.

Attribute Macros

request

⚠ If this is the only documentation you see, please navigate to the docs for ruma_common::api::request, where actual documentation can be found.

response

⚠ If this is the only documentation you see, please navigate to the docs for ruma_common::api::response, where actual documentation can be found.