Module matrix_sdk_base::ruma::api::client::discovery::get_capabilities

GET /_matrix/client/*/capabilities

Get information about the server’s supported feature set and other relevant capabilities (spec).

Modules

iter

Iterator implementation for Capabilities

v3

/v3/ (spec)

Structs

Capabilities

Contains information about all the capabilities that the server supports.

ChangePasswordCapability

Information about the m.change_password capability

RoomVersionsCapability

Information about the m.room_versions capability

SetAvatarUrlCapability

Information about the m.set_avatar_url capability

SetDisplayNameCapability

Information about the m.set_displayname capability

ThirdPartyIdChangesCapability

Information about the m.3pid_changes capability

Enums

RoomVersionStability

The stability of a room version.