Module matrix_sdk_base::ruma::api::client::search::search_events::v3

/v3/ (spec)

Structs

Categories

Categories of events that can be searched for.

Criteria

Criteria for searching a category of events.

EventContext

Configures whether any context for the events returned are included in the response.

EventContextResult

Context for search results, if requested.

Grouping

A grouping for partitioning the result set.

Groupings

Requests that the server partitions the result set based on the provided list of keys.

Request

Request type for the search endpoint.

Response

Response type for the search endpoint.

ResultCategories

Categories of events that can be searched for.

ResultGroup

A grouping of results, if requested.

ResultRoomEvents

Categories of events that can be searched for.

SearchResult

A search result.

UserProfile

A user profile.

Enums

GroupingKey

The key within events to use for this grouping.

OrderBy

The order in which to search for results.

OwnedRoomIdOrUserId

Represents either a room or user ID for returning grouped search results.

SearchKeys

The keys to search for.