Module matrix_sdk_base::ruma::api::client::sync::sync_events::v4

POST /_matrix/client/unstable/org.matrix.msc3575/sync (MSC)

Get all new events in a sliding window of rooms since the last sync or a given point in time.

Structs

AccountData

Account-data extension response data.

AccountDataConfig

Account-data extension configuration.

E2EE

E2EE extension response data.

E2EEConfig

E2EE extension configuration.

Extensions

Extensions specific response data.

ExtensionsConfig

Sliding-Sync extension configuration.

IncludeOldRooms

Configuration for old rooms to include

Receipts

Receipt extension response data.

ReceiptsConfig

Receipt extension configuration.

Request

Request type for the sync endpoint.

Response

Response type for the sync endpoint.

RoomDetailsConfig

Configuration for requesting room details.

RoomSubscription

Configuration for room subscription

SlidingSyncRoom

Updates to joined rooms.

SyncList

Updates to joined rooms.

SyncOp

Updates to joined rooms.

SyncRequestList

Sliding Sync Request for each list.

SyncRequestListFilters

Filter for a sliding sync list, set at request.

ToDevice

To-device messages extension response.

ToDeviceConfig

To-device messages extension configuration.

Typing

Typing extension response data.

TypingConfig

Typing extension configuration.

Enums

RoomReceiptConfig

Single entry for a room-related read receipt configuration in ReceiptsConfig.

SlidingOp

Operation applied to the specific SlidingSyncList