Enum matrix_sdk_base::ruma::api::client::uiaa::AuthType

#[non_exhaustive]
pub enum AuthType {
    Password,
    ReCaptcha,
    EmailIdentity,
    Msisdn,
    Sso,
    Dummy,
    RegistrationToken,
}

The type of an authentication stage.

This type can hold an arbitrary string. To build this with a custom value, convert it from a string with ::from() / .into(). To check for values that are not available as a documented variant here, use its string representation, obtained through .as_str().

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Password

Password-based authentication (m.login.password).

ReCaptcha

Google ReCaptcha 2.0 authentication (m.login.recaptcha).

EmailIdentity

Email-based authentication (m.login.email.identity).

Msisdn

Phone number-based authentication (m.login.msisdn).

Sso

SSO-based authentication (m.login.sso).

Dummy

Dummy authentication (m.login.dummy).

RegistrationToken

Registration token-based authentication (m.login.registration_token).

Implementations

impl AuthType

pub fn as_str(&self) -> &str

Creates a string slice from this AuthType.