Crate matrix_sdk_base::ruma::api::federation

This is supported on crate feature api only.

(De)serializable types for the Matrix Server-Server API. These types are used by server code.

Modules

authorization

Endpoints to retrieve the complete auth chain for a given event.

backfill

Endpoints to request more history from another homeserver.

device

Endpoints to retrieve information about user devices

directory

Room directory endpoints.

discovery

Server discovery endpoints.

event

Endpoints to get general information about events

keys

Endpoints for handling keys for end-to-end encryption

knock

Endpoints for handling room knocking.

membership

Room membership endpoints.

openid

OpenID endpoints.

query

Endpoints to retrieve information from a homeserver about a resource.

thirdparty

Module for dealing with third party identifiers

transactions

Endpoints for exchanging transaction messages between homeservers.