Trait matrix_sdk_base::ruma::api::IncomingRequest

pub trait IncomingRequest: Sized {
    type EndpointError: EndpointError;
    type OutgoingResponse: OutgoingResponse;

    const METADATA: Metadata;

    // Required method
    fn try_from_http_request<B, S>(
        req: Request<B>,
        path_args: &[S]
    ) -> Result<Self, FromHttpRequestError>
       where B: AsRef<[u8]>,
             S: AsRef<str>;
}

A request type for a Matrix API endpoint, used for receiving requests.

Required Associated Types

type EndpointError: EndpointError

A type capturing the error conditions that can be returned in the response.

type OutgoingResponse: OutgoingResponse

Response type to return when the request is successful.

Required Associated Constants

const METADATA: Metadata

Metadata about the endpoint.

Required Methods

fn try_from_http_request<B, S>( req: Request<B>, path_args: &[S] ) -> Result<Self, FromHttpRequestError>
where B: AsRef<[u8]>, S: AsRef<str>,

Tries to turn the given http::Request into this request type, together with the corresponding path arguments.

Note: The strings in path_args need to be percent-decoded.

Object Safety

This trait is not object safe.