Type Alias matrix_sdk_base::ruma::events::policy::rule::room::InitialPolicyRuleRoomEvent

pub type InitialPolicyRuleRoomEvent = InitialStateEvent<PolicyRuleRoomEventContent>;

An m.policy.rule.room event for creating a room.

Aliased Type

struct InitialPolicyRuleRoomEvent {
    pub content: PolicyRuleRoomEventContent,
    pub state_key: String,
}

Fields
content: PolicyRuleRoomEventContent

Data specific to the event type.

state_key: String

A unique key which defines the overwriting semantics for this piece of room state.

This is often an empty string, but some events send a UserId to show which user the event affects.

Defaults to the empty string.