Type Alias matrix_sdk_base::ruma::events::poll::unstable_end::RedactedSyncUnstablePollEndEvent

pub type RedactedSyncUnstablePollEndEvent = RedactedSyncMessageLikeEvent<RedactedUnstablePollEndEventContent>;

An org.matrix.msc3381.poll.end event from a sync_events response that has been redacted.

Aliased Type

struct RedactedSyncUnstablePollEndEvent {
    pub content: RedactedUnstablePollEndEventContent,
    pub event_id: OwnedEventId,
    pub sender: OwnedUserId,
    pub origin_server_ts: MilliSecondsSinceUnixEpoch,
    pub unsigned: RedactedUnsigned,
}

Fields
content: RedactedUnstablePollEndEventContent

Data specific to the event type.

event_id: OwnedEventId

The globally unique event identifier for the user who sent the event.

sender: OwnedUserId

The fully-qualified ID of the user who sent this event.

origin_server_ts: MilliSecondsSinceUnixEpoch

Timestamp in milliseconds on originating homeserver when this event was sent.

unsigned: RedactedUnsigned

Additional key-value pairs not signed by the homeserver.