Type Alias matrix_sdk_base::ruma::events::room::member::PossiblyRedactedRoomMemberEventContent

pub type PossiblyRedactedRoomMemberEventContent = RoomMemberEventContent;

The possibly redacted form of RoomMemberEventContent.

This type is used when it’s not obvious whether the content is redacted or not.

Aliased Type

struct PossiblyRedactedRoomMemberEventContent {
    pub avatar_url: Option<OwnedMxcUri>,
    pub displayname: Option<String>,
    pub is_direct: Option<bool>,
    pub membership: MembershipState,
    pub third_party_invite: Option<ThirdPartyInvite>,
    pub blurhash: Option<String>,
    pub reason: Option<String>,
    pub join_authorized_via_users_server: Option<OwnedUserId>,
}

Fields
avatar_url: Option<OwnedMxcUri>

The avatar URL for this user, if any.

This is added by the homeserver. If you activate the compat-empty-string-null feature, this field being an empty string in JSON will result in None here during deserialization.

displayname: Option<String>

The display name for this user, if any.

This is added by the homeserver.

is_direct: Option<bool>

Flag indicating whether the room containing this event was created with the intention of being a direct chat.

membership: MembershipState

The membership state of this user.

third_party_invite: Option<ThirdPartyInvite>

If this member event is the successor to a third party invitation, this field will contain information about that invitation.

blurhash: Option<String>

The BlurHash for the avatar pointed to by avatar_url.

This uses the unstable prefix in MSC2448.

reason: Option<String>

User-supplied text for why their membership has changed.

For kicks and bans, this is typically the reason for the kick or ban. For other membership changes, this is a way for the user to communicate their intent without having to send a message to the room, such as in a case where Bob rejects an invite from Alice about an upcoming concert, but can’t make it that day.

Clients are not recommended to show this reason to users when receiving an invite due to the potential for spam and abuse. Hiding the reason behind a button or other component is recommended.

join_authorized_via_users_server: Option<OwnedUserId>

Arbitrarily chosen UserId (MxID) of a local user who can send an invite.