Type Alias matrix_sdk_base::ruma::events::room::server_acl::OriginalRoomServerAclEvent

pub type OriginalRoomServerAclEvent = OriginalStateEvent<RoomServerAclEventContent>;

An m.room.server_acl event.

Aliased Type

struct OriginalRoomServerAclEvent {
    pub content: RoomServerAclEventContent,
    pub event_id: OwnedEventId,
    pub sender: OwnedUserId,
    pub origin_server_ts: MilliSecondsSinceUnixEpoch,
    pub room_id: OwnedRoomId,
    pub state_key: EmptyStateKey,
    pub unsigned: StateUnsigned<RoomServerAclEventContent>,
}

Fields
content: RoomServerAclEventContent

Data specific to the event type.

event_id: OwnedEventId

The globally unique event identifier for the user who sent the event.

sender: OwnedUserId

The fully-qualified ID of the user who sent this event.

origin_server_ts: MilliSecondsSinceUnixEpoch

Timestamp in milliseconds on originating homeserver when this event was sent.

room_id: OwnedRoomId

The ID of the room associated with this event.

state_key: EmptyStateKey

A unique key which defines the overwriting semantics for this piece of room state.

This is often an empty string, but some events send a UserId to show which user the event affects.

unsigned: StateUnsigned<RoomServerAclEventContent>

Additional key-value pairs not signed by the homeserver.