Struct matrix_sdk_base::store::StateChanges

pub struct StateChanges {
    pub sync_token: Option<String>,
    pub account_data: BTreeMap<GlobalAccountDataEventType, Raw<AnyGlobalAccountDataEvent>>,
    pub presence: BTreeMap<OwnedUserId, Raw<PresenceEvent>>,
    pub profiles: BTreeMap<OwnedRoomId, BTreeMap<OwnedUserId, MinimalRoomMemberEvent>>,
    pub state: BTreeMap<OwnedRoomId, BTreeMap<StateEventType, BTreeMap<String, Raw<AnySyncStateEvent>>>>,
    pub room_account_data: BTreeMap<OwnedRoomId, BTreeMap<RoomAccountDataEventType, Raw<AnyRoomAccountDataEvent>>>,
    pub room_infos: BTreeMap<OwnedRoomId, RoomInfo>,
    pub receipts: BTreeMap<OwnedRoomId, ReceiptEventContent>,
    pub redactions: BTreeMap<OwnedRoomId, BTreeMap<OwnedEventId, Raw<SyncRoomRedactionEvent>>>,
    pub stripped_state: BTreeMap<OwnedRoomId, BTreeMap<StateEventType, BTreeMap<String, Raw<AnyStrippedStateEvent>>>>,
    pub ambiguity_maps: BTreeMap<OwnedRoomId, BTreeMap<String, BTreeSet<OwnedUserId>>>,
}

Store state changes and pass them to the StateStore.

Fields
sync_token: Option<String>

The sync token that relates to this update.

account_data: BTreeMap<GlobalAccountDataEventType, Raw<AnyGlobalAccountDataEvent>>

A mapping of event type string to AnyBasicEvent.

presence: BTreeMap<OwnedUserId, Raw<PresenceEvent>>

A mapping of UserId to PresenceEvent.

profiles: BTreeMap<OwnedRoomId, BTreeMap<OwnedUserId, MinimalRoomMemberEvent>>

A mapping of RoomId to a map of users and their MinimalRoomMemberEvent.

state: BTreeMap<OwnedRoomId, BTreeMap<StateEventType, BTreeMap<String, Raw<AnySyncStateEvent>>>>

A mapping of RoomId to a map of event type string to a state key and AnySyncStateEvent.

room_account_data: BTreeMap<OwnedRoomId, BTreeMap<RoomAccountDataEventType, Raw<AnyRoomAccountDataEvent>>>

A mapping of RoomId to a map of event type string to AnyBasicEvent.

room_infos: BTreeMap<OwnedRoomId, RoomInfo>

A map of RoomId to RoomInfo.

receipts: BTreeMap<OwnedRoomId, ReceiptEventContent>

A map of RoomId to ReceiptEventContent.

redactions: BTreeMap<OwnedRoomId, BTreeMap<OwnedEventId, Raw<SyncRoomRedactionEvent>>>

A map of RoomId to maps of OwnedEventId to be redacted by SyncRoomRedactionEvent.

stripped_state: BTreeMap<OwnedRoomId, BTreeMap<StateEventType, BTreeMap<String, Raw<AnyStrippedStateEvent>>>>

A mapping of RoomId to a map of event type to a map of state key to AnyStrippedStateEvent.

ambiguity_maps: BTreeMap<OwnedRoomId, BTreeMap<String, BTreeSet<OwnedUserId>>>

A map from room id to a map of a display name and a set of user ids that share that display name in the given room.

Implementations

impl StateChanges

pub fn new(sync_token: String) -> Self

Create a new StateChanges struct with the given sync_token.

pub fn add_presence_event( &mut self, event: PresenceEvent, raw_event: Raw<PresenceEvent> )

Update the StateChanges struct with the given PresenceEvent.

pub fn add_room(&mut self, room: RoomInfo)

Update the StateChanges struct with the given RoomInfo.

pub fn add_account_data( &mut self, event: AnyGlobalAccountDataEvent, raw_event: Raw<AnyGlobalAccountDataEvent> )

Update the StateChanges struct with the given AnyBasicEvent.

pub fn add_room_account_data( &mut self, room_id: &RoomId, event: AnyRoomAccountDataEvent, raw_event: Raw<AnyRoomAccountDataEvent> )

Update the StateChanges struct with the given room with a new AnyBasicEvent.

pub fn add_stripped_member( &mut self, room_id: &RoomId, user_id: &UserId, event: Raw<StrippedRoomMemberEvent> )

Update the StateChanges struct with the given room with a new StrippedMemberEvent.

pub fn add_state_event( &mut self, room_id: &RoomId, event: AnySyncStateEvent, raw_event: Raw<AnySyncStateEvent> )

Update the StateChanges struct with the given room with a new AnySyncStateEvent.

pub fn add_redaction( &mut self, room_id: &RoomId, redacted_event_id: &EventId, redaction: Raw<SyncRoomRedactionEvent> )

Redact an event in the room

pub fn add_receipts(&mut self, room_id: &RoomId, event: ReceiptEventContent)

Update the StateChanges struct with the given room with a new Receipts.