Trait matrix_sdk_base::store::StateStoreExt

pub trait StateStoreExt: StateStore {
    fn get_state_event_static<'life0, 'life1, 'async_trait, C>(
        &'life0 self,
        room_id: &'life1 RoomId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<SyncStateEvent<C>>>>> + Send + 'async_trait>>
    where
        C: StaticEventContent + StateEventContent<StateKey = EmptyStateKey> + RedactContent,
        C::Redacted: RedactedStateEventContent + DeserializeOwned,
        C: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn get_state_event_static_for_key<'life0, 'life1, 'life2, 'async_trait, C, K>(
        &'life0 self,
        room_id: &'life1 RoomId,
        state_key: &'life2 K
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<SyncStateEvent<C>>>>> + Send + 'async_trait>>
    where
        C: StaticEventContent + StateEventContent + RedactContent,
        C::StateKey: Borrow<K>,
        C::Redacted: RedactedStateEventContent,
        K: AsRef<str> + ?Sized + Sync,
        C: 'async_trait,
        K: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn get_state_events_static<'life0, 'life1, 'async_trait, C>(
        &'life0 self,
        room_id: &'life1 RoomId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Raw<SyncStateEvent<C>>>>> + Send + 'async_trait>>
    where
        C: StaticEventContent + StateEventContent + RedactContent,
        C::Redacted: RedactedStateEventContent,
        C: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn get_account_data_event_static<'life0, 'async_trait, C>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<GlobalAccountDataEvent<C>>>>> + Send + 'async_trait>>
    where
        C: StaticEventContent + GlobalAccountDataEventContent,
        C: 'async_trait,
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn get_room_account_data_event_static<'life0, 'life1, 'async_trait, C>(
        &'life0 self,
        room_id: &'life1 RoomId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<RoomAccountDataEvent<C>>>>> + Send + 'async_trait>>
    where
        C: StaticEventContent + RoomAccountDataEventContent,
        C: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Convenience functionality for state stores.

Provided Methods

fn get_state_event_static<'life0, 'life1, 'async_trait, C>(
&'life0 self,
room_id: &'life1 RoomId
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<SyncStateEvent<C>>>>> + Send + 'async_trait>>where
C: StaticEventContent + StateEventContent<StateKey = EmptyStateKey> + RedactContent,
C::Redacted: RedactedStateEventContent + DeserializeOwned,
C: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Get a specific state event of statically-known type.

Arguments

room_id - The id of the room the state event was received for.

fn get_state_event_static_for_key<'life0, 'life1, 'life2, 'async_trait, C, K>(
&'life0 self,
room_id: &'life1 RoomId,
state_key: &'life2 K
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<SyncStateEvent<C>>>>> + Send + 'async_trait>>where
C: StaticEventContent + StateEventContent + RedactContent,
C::StateKey: Borrow<K>,
C::Redacted: RedactedStateEventContent,
K: AsRef<str> + ?Sized + Sync,
C: 'async_trait,
K: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: Sync + 'async_trait,

Get a specific state event of statically-known type.

Arguments

room_id - The id of the room the state event was received for.

fn get_state_events_static<'life0, 'life1, 'async_trait, C>(
&'life0 self,
room_id: &'life1 RoomId
) -> Pin<Box<dyn Future<Output = Result<Vec<Raw<SyncStateEvent<C>>>>> + Send + 'async_trait>>where
C: StaticEventContent + StateEventContent + RedactContent,
C::Redacted: RedactedStateEventContent,
C: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Get a list of state events of a statically-known type for a given room.

Arguments

room_id - The id of the room to find events for.

fn get_account_data_event_static<'life0, 'async_trait, C>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<GlobalAccountDataEvent<C>>>>> + Send + 'async_trait>>where
C: StaticEventContent + GlobalAccountDataEventContent,
C: 'async_trait,
'life0: 'async_trait,
Self: Sync + 'async_trait,

Get an event of a statically-known type from the account data store.

fn get_room_account_data_event_static<'life0, 'life1, 'async_trait, C>(
&'life0 self,
room_id: &'life1 RoomId
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<RoomAccountDataEvent<C>>>>> + Send + 'async_trait>>where
C: StaticEventContent + RoomAccountDataEventContent,
C: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Get an event of a statically-known type from the room account data store.

Arguments

room_id - The id of the room for which the room account data event should be fetched.