Struct matrix_sdk_base::StateChanges

pub struct StateChanges {

    pub sync_token: Option<String>,
    pub session: Option<Session>,
    pub account_data: BTreeMap<String, Raw<AnyBasicEvent>>,
    pub presence: BTreeMap<UserId, Raw<PresenceEvent>>,
    pub members: BTreeMap<RoomId, BTreeMap<UserId, MemberEvent>>,
    pub profiles: BTreeMap<RoomId, BTreeMap<UserId, MemberEventContent>>,
    pub state: BTreeMap<RoomId, BTreeMap<String, BTreeMap<String, Raw<AnySyncStateEvent>>>>,
    pub room_account_data: BTreeMap<RoomId, BTreeMap<String, Raw<AnyBasicEvent>>>,
    pub room_infos: BTreeMap<RoomId, RoomInfo>,
    pub stripped_state: BTreeMap<RoomId, BTreeMap<String, BTreeMap<String, Raw<AnyStrippedStateEvent>>>>,
    pub stripped_members: BTreeMap<RoomId, BTreeMap<UserId, StrippedMemberEvent>>,
    pub invited_room_info: BTreeMap<RoomId, RoomInfo>,
    pub ambiguity_maps: BTreeMap<RoomId, BTreeMap<String, BTreeSet<UserId>>>,
    pub notifications: BTreeMap<RoomId, Vec<Notification>>,

}

Store state changes and pass them to the StateStore.

Fields
sync_token: Option<String>

The sync token that relates to this update.

session: Option<Session>

A user session, containing an access token and information about the associated user account.

account_data: BTreeMap<String, Raw<AnyBasicEvent>>

A mapping of event type string to AnyBasicEvent.

presence: BTreeMap<UserId, Raw<PresenceEvent>>

A mapping of UserId to PresenceEvent.

members: BTreeMap<RoomId, BTreeMap<UserId, MemberEvent>>

A mapping of RoomId to a map of users and their MemberEvent.

profiles: BTreeMap<RoomId, BTreeMap<UserId, MemberEventContent>>

A mapping of RoomId to a map of users and their MemberEventContent.

state: BTreeMap<RoomId, BTreeMap<String, BTreeMap<String, Raw<AnySyncStateEvent>>>>

A mapping of RoomId to a map of event type string to a state key and AnySyncStateEvent.

room_account_data: BTreeMap<RoomId, BTreeMap<String, Raw<AnyBasicEvent>>>

A mapping of RoomId to a map of event type string to AnyBasicEvent.

room_infos: BTreeMap<RoomId, RoomInfo>

A map of RoomId to RoomInfo.

stripped_state: BTreeMap<RoomId, BTreeMap<String, BTreeMap<String, Raw<AnyStrippedStateEvent>>>>

A mapping of RoomId to a map of event type to a map of state key to AnyStrippedStateEvent.

stripped_members: BTreeMap<RoomId, BTreeMap<UserId, StrippedMemberEvent>>

A mapping of RoomId to a map of users and their StrippedMemberEvent.

invited_room_info: BTreeMap<RoomId, RoomInfo>

A map of RoomId to RoomInfo.

ambiguity_maps: BTreeMap<RoomId, BTreeMap<String, BTreeSet<UserId>>>

A map from room id to a map of a display name and a set of user ids that share that display name in the given room.

notifications: BTreeMap<RoomId, Vec<Notification>>

A map of RoomId to a vector of Notifications

Implementations

impl StateChanges

pub fn new(sync_token: String) -> Self

Create a new StateChanges struct with the given sync_token.

pub fn add_presence_event(
&mut self,
event: PresenceEvent,
raw_event: Raw<PresenceEvent>
)

Update the StateChanges struct with the given PresenceEvent.

pub fn add_room(&mut self, room: RoomInfo)

Update the StateChanges struct with the given RoomInfo.

pub fn add_stripped_room(&mut self, room: RoomInfo)

Update the StateChanges struct with the given RoomInfo.

pub fn add_account_data(
&mut self,
event: AnyBasicEvent,
raw_event: Raw<AnyBasicEvent>
)

Update the StateChanges struct with the given AnyBasicEvent.

pub fn add_room_account_data(
&mut self,
room_id: &RoomId,
event: AnyBasicEvent,
raw_event: Raw<AnyBasicEvent>
)

Update the StateChanges struct with the given room with a new AnyBasicEvent.

pub fn add_stripped_member(
&mut self,
room_id: &RoomId,
event: StrippedMemberEvent
)

Update the StateChanges struct with the given room with a new StrippedMemberEvent.

pub fn add_state_event(
&mut self,
room_id: &RoomId,
event: AnySyncStateEvent,
raw_event: Raw<AnySyncStateEvent>
)

Update the StateChanges struct with the given room with a new AnySyncStateEvent.

pub fn add_notification(&mut self, room_id: &RoomId, notification: Notification)

Update the StateChanges struct with the given room with a new Notification.