Module matrix_sdk_base::thirdparty

Common types for the third party networks module.

Structs

FieldType
A type definition for a field used to identify third party users or locations.

FieldTypeInit
Initial set of fields of FieldType.

Location
A third party network location.

Protocol
Metadata about a third party protocol.

ProtocolInit
Initial set of fields of Protocol.

ProtocolInstance
Metadata about an instance of a third party protocol.

ProtocolInstanceInit
Initial set of fields of Protocol.

User
A third party network user.

Enums

Medium
The medium of a third party identifier.