Trait matrix_sdk_base::IncomingRequest

pub trait IncomingRequest {
    type EndpointError: EndpointError;
    type OutgoingResponse: OutgoingResponse;

    const METADATA: Metadata;

    fn try_from_http_request<T>(
        req: Request<T>
    ) -> Result<Self, FromHttpRequestError>
    where
        T: AsRef<[u8]>;
}

A request type for a Matrix API endpoint, used for receiving requests.

Associated Types

type EndpointError: EndpointError

A type capturing the error conditions that can be returned in the response.

type OutgoingResponse: OutgoingResponse

Response type to return when the request is successful.

Loading content...
Associated Constants

const METADATA: Metadata

Metadata about the endpoint.

Loading content...
Required methods

fn try_from_http_request<T>(
req: Request<T>
) -> Result<Self, FromHttpRequestError> where
T: AsRef<[u8]>,

Tries to turn the given http::Request into this request type.

Loading content...