Crate matrix_sdk_base

This crate implements the base to build a Matrix client library.

Crate Feature Flags

The following crate feature flags are available:

encryption: Enables end-to-end encryption support in the library.

qrcode: Enables QRcode generation and reading code.

testing: Provides facilities and functions for tests, in particular for integration testing store implementations. ATTENTION: do not ever use outside of tests, we do not provide any stability warantees on these, these are merely helpers. If you find you need any function provided here outside of tests, please open a Github Issue and inform us about your use case for us to consider.

Re-exports

pub use matrix_sdk_crypto as crypto;

pub use once_cell;

pub use store::StateChanges;

pub use store::StateStore;

pub use store::Store;

pub use store::StoreError;

Modules

deserialized_responses

executor

Abstraction over an executor so we can spawn tasks under WASM the same way we do usually.

instant

locks

media

Common types for media content.

store

The state store holds the overall state for rooms, users and their profiles and their timelines. It is an overall cache for faster access and convenience- accessible through Store.

Structs

BaseClient

A no IO Client implementation.

OriginalMinimalStateEvent

An unredacted minimal state event.

RedactedMinimalStateEvent

A redacted minimal state event.

Room

The underlying room data structure collecting state for joined, left and invited rooms.

RoomInfo

The underlying pure data structure for joined and left rooms.

RoomMember

A member of a room.

Session

A user session, containing an access token and information about the associated user account.

Enums

DisplayName

The name of the room, either from the metadata or calculated according to matrix specification

Error

Internal representation of errors.

MinimalStateEvent

A minimal state event.

RoomType

Enum keeping track in which state the room is, e.g. if our own user is joined, invited, or has left the room.

Traits

AsyncTraitDeps

Super trait that is used for our store traits, this trait will differ if it’s used on WASM. WASM targets will not require Send and Sync to have implemented, while other targets will.

Type Definitions

MinimalRoomMemberEvent

A minimal m.room.member event.

Result

Result type of the rust-sdk.