Crate matrix_sdk_base

This crate implements a Matrix client library.

Crate Feature Flags

The following crate feature flags are available:

encryption: Enables end-to-end encryption support in the library.

sled_cryptostore: Enables a Sled based store for the encryption keys. If this is disabled and encryption support is enabled the keys will by default be stored only in memory and thus lost after the client is destroyed.

markdown: Support for sending markdown formatted messages.

Re-exports

pub use matrix_sdk_crypto as crypto;

Modules

deserialized_responses

executor
Abstraction over an executor so we can spawn tasks under WASM the same way we do usually.

instant

locks

media
Common types for media content.

uuid
Generate and parse UUIDs.

Structs

BaseClient
A no IO Client implementation.

BaseClientConfig
Configuration for the creation of the BaseClient.

Room
The underlying room data structure collecting state for joined, left and invited rooms.

RoomInfo
The underlying pure data structure for joined and left rooms.

RoomMember
A member of a room.

Session
A user session, containing an access token and information about the associated user account.

StateChanges
Store state changes and pass them to the StateStore.

Store
A state store wrapper for the SDK.

Enums

Error
Internal representation of errors.

RoomType
Enum keeping track in which state the room is, e.g. if our own user is joined, invited, or has left the room.

StoreError
State store specific error type.

Traits

AsyncTraitDeps
Super trait that is used for our store traits, this trait will differ if it’s used on WASM. WASM targets will not require Send and Sync to have implemented, while other targets will.

StateStore
An abstract state store trait that can be used to implement different stores for the SDK.

Type Definitions

Result
Result type of the rust-sdk.

Attribute Macros

async_trait