Crate matrix_sdk_base

This crate implements the base to build a Matrix client library.

Crate Feature Flags

The following crate feature flags are available:

encryption: Enables end-to-end encryption support in the library.

qrcode: Enbles QRcode generation and reading code

store_key: Provides extra facilities for StateStores to manage store keys

testing: provides facilities and functions for tests, in particular for integration testing store implementations. ATTENTION: do not ever use outside of tests, we do not provide any stability warantees on these, these are merely helpers. If you find you need any function provided here outside of tests, please open a Github Issue and inform us about your use case for us to consider.

Re-exports

pub use http;

pub use matrix_sdk_crypto as crypto;

pub use store::StateChanges;

pub use store::StateStore;

pub use store::Store;

pub use store::StoreError;

Modules

deserialized_responses

executor

Abstraction over an executor so we can spawn tasks under WASM the same way we do usually.

instant

locks

media

Common types for media content.

ruma

Types and traits for working with the Matrix protocol.

store

The state store holds the overall state for rooms, users and their profiles and their timelines. It is an overall cache for faster access and convenience- accessible through Store.

util

Macros

statestore_integration_tests

Macro building to allow your StateStore implementation to run the entire tests suite locally.

Structs

BaseClient

A no IO Client implementation.

Room

The underlying room data structure collecting state for joined, left and invited rooms.

RoomInfo

The underlying pure data structure for joined and left rooms.

RoomMember

A member of a room.

Session

A user session, containing an access token and information about the associated user account.

Enums

Error

Internal representation of errors.

RoomType

Enum keeping track in which state the room is, e.g. if our own user is joined, invited, or has left the room.

TimelineStreamError

Errors in a timeline stream

Traits

AsyncTraitDepsNon-WebAssembly

Super trait that is used for our store traits, this trait will differ if it’s used on WASM. WASM targets will not require Send and Sync to have implemented, while other targets will.

Type Definitions

Result

Result type of the rust-sdk.

Attribute Macros

async_trait