Trait matrix_sdk_base::media::UniqueKey

pub trait UniqueKey {
    fn unique_key(&self) -> String;
}

A trait to uniquely identify values of the same type.

Required Methods

fn unique_key(&self) -> String

A string that uniquely identifies Self compared to other values of the same type.