Module matrix_sdk_base::store

The state store holds the overall state for rooms, users and their profiles and their timelines. It is an overall cache for faster access and convenience- accessible through Store.

Implementing the StateStore trait, you can plug any storage backend into the store for the actual storage. By default this brings an in-memory store.

Modules

migration_helpers

Data migration helpers for StateStore implementations.

Structs

MemoryStore

In-Memory, non-persistent implementation of the StateStore

StateChanges

Store state changes and pass them to the StateStore.

StoreConfig

Configuration for the state store and, when encryption is enabled, for the crypto store.

Enums

StateStoreDataKey

A key for key-value data.

StateStoreDataValue

A value for key-value data that should be persisted into the store.

StoreEncryptionError

Error type for the StoreCipher operations.

StoreError

State store specific error type.

Traits

IntoStateStore

A type that can be type-erased into Arc<dyn StateStore>.

StateStore

An abstract state store trait that can be used to implement different stores for the SDK.

StateStoreExt

Convenience functionality for state stores.

Type Aliases

BoxStream

BoxStream of owned Types

DynStateStore

A type-erased StateStore.

Result

A StateStore specific result type.