Struct matrix_sdk_base::BaseClient

pub struct BaseClient { /* private fields */ }

A no IO Client implementation.

This Client is a state machine that receives responses and events and accordingly updates its state.

Implementations

impl BaseClient

pub fn new() -> Self

Create a new default client.

pub fn with_store_config(config: StoreConfig) -> Self

Create a new client.

Arguments

config - An optional session if the user already has one from a previous login call.

pub fn session_meta(&self) -> Option<&SessionMeta>

Get the session meta information.

If the client is currently logged in, this will return a SessionMeta object which contains the user ID and device ID. Otherwise it returns None.

pub fn session_tokens(&self) -> Subscriber<Option<SessionTokens>>

Get the session tokens.

This returns a subscriber object that you can use both to get the current value as well as to react to changes to the tokens.

If the client is currently logged in, the inner value is a SessionTokens object which contains the access token and optional refresh token. Otherwise it is None.

pub fn set_session_tokens(&self, tokens: SessionTokens)

Set the session tokens.

pub fn session(&self) -> Option<Session>

Get the user login session.

If the client is currently logged in, this will return a Session object.

Returns a session object if the client is logged in. Otherwise returns None.

pub fn get_rooms(&self) -> Vec<Room>

Get all the rooms this client knows about.

pub async fn get_or_create_room( &self, room_id: &RoomId, room_state: RoomState ) -> Room

Lookup the Room for the given RoomId, or create one, if it didn’t exist yet in the store

pub fn get_stripped_rooms(&self) -> Vec<Room>

Get all the rooms this client knows about.

pub fn store(&self) -> &DynStateStore

Get a reference to the store.

pub fn logged_in(&self) -> bool

Is the client logged in.

pub async fn receive_login_response(&self, response: &Response) -> Result<()>

Receive a login response and update the session of the client.

Arguments

response - A successful login response that contains our access token and device ID.

pub async fn set_session_meta(&self, session_meta: SessionMeta) -> Result<()>

Set the meta of the session.

If encryption is enabled, this also initializes or restores the OlmMachine.

Arguments

session_meta - The meta of a session that the user already has from a previous login call.

This method panics if it is called twice.

pub async fn sync_token(&self) -> Option<String>

Get the current, if any, sync token of the client. This will be None if the client didn’t sync at least once.

pub async fn room_joined(&self, room_id: &RoomId) -> Result<Room>

User has joined a room.

Update the internal and cached state accordingly. Return the final Room.

pub async fn room_left(&self, room_id: &RoomId) -> Result<Room>

User has left a room.

Update the internal and cached state accordingly. Return the final Room.

pub fn sync_lock(&self) -> &RwLock<()>

Get access to the store’s sync lock.

pub async fn receive_sync_response( &self, response: Response ) -> Result<SyncResponse>

Receive a response from a sync call.

Arguments

response - The response that we received after a successful sync.

pub async fn receive_members( &self, room_id: &RoomId, response: &Response ) -> Result<MembersResponse>

Receive a get member events response and convert it to a deserialized MembersResponse

Arguments

room_id - The room id this response belongs to.

response - The raw response that was received from the server.

pub async fn receive_filter_upload( &self, filter_name: &str, response: &Response ) -> Result<()>

Receive a successful filter upload response, the filter id will be stored under the given name in the store.

The filter id can later be retrieved with the get_filter method.

Arguments

filter_name - The name that should be used to persist the filter id in the store.

response - The successful filter upload response containing the filter id.

pub async fn get_filter(&self, filter_name: &str) -> StoreResult<Option<String>>

Get the filter id of a previously uploaded filter.

Note: A filter will first need to be uploaded and persisted using receive_filter_upload.

Arguments

filter_name - The name of the filter that was previously used to persist the filter.

pub async fn share_room_key( &self, room_id: &RoomId ) -> Result<Vec<Arc<ToDeviceRequest>>>

Available on crate feature e2e-encryption only.

Get a to-device request that will share a room key with users in a room.

pub fn get_room(&self, room_id: &RoomId) -> Option<Room>

Get the room with the given room id.

Arguments

room_id - The id of the room that should be fetched.

pub fn olm_machine(&self) -> Option<&OlmMachine>

Available on crate feature e2e-encryption only.

Get the olm machine.

pub async fn get_push_rules(&self, changes: &StateChanges) -> Result<Ruleset>

Get the push rules.

Gets the push rules from changes if they have been updated, otherwise get them from the store. As a fallback, uses Ruleset::server_default if the user is logged in.

pub async fn get_push_room_context( &self, room: &Room, room_info: &RoomInfo, changes: &StateChanges ) -> Result<Option<PushConditionRoomCtx>>

Get the push context for the given room.

Tries to get the data from changes or the up to date room_info. Loads the data from the store otherwise.

Returns None if some data couldn’t be found. This should only happen in brand new rooms, while we process its state.

pub async fn update_push_room_context( &self, push_rules: &mut PushConditionRoomCtx, user_id: &UserId, room_info: &RoomInfo, changes: &StateChanges )

Update the push context for the given room.

Updates the context data from changes or room_info.

pub fn subscribe_to_ignore_user_list_changes(&self) -> Subscriber<()>

Returns a subscriber that publishes an event every time the ignore user list changes