Struct matrix_sdk_base::Room

pub struct Room {
    pub latest_encrypted_events: Arc<RwLock<RingBuffer<Raw<AnySyncTimelineEvent>>>>,
    /* private fields */
}

The underlying room data structure collecting state for joined, left and invited rooms.

Fields
latest_encrypted_events: Arc<RwLock<RingBuffer<Raw<AnySyncTimelineEvent>>>>

The most recent few encrypted events. When the keys come through to decrypt these, the most recent relevant one will replace latest_event. (We can’t tell which one is relevant until they are decrypted.)

Currently, these are held in Room rather than RoomInfo, because we were not sure whether holding too many of them might make the cache too slow to load on startup. Keeping them here means they are not cached to disk but held in memory.

Implementations

impl Room

pub fn room_id(&self) -> &RoomId

Get the unique room id of the room.

pub fn own_user_id(&self) -> &UserId

Get our own user id.

pub fn state(&self) -> RoomState

Get the state of the room.

pub fn is_space(&self) -> bool

Whether this room’s RoomType is m.space.

pub fn unread_notification_counts(&self) -> UnreadNotificationsCount

Get the unread notification counts.

pub fn num_unread_messages(&self) -> u64

Get the number of unread messages (computed client-side).

This might be more precise than Self::unread_notification_counts for encrypted rooms.

pub fn read_receipts(&self) -> RoomReadReceipts

Get the detailed information about read receipts for the room.

pub fn num_unread_notifications(&self) -> u64

Get the number of unread notifications (computed client-side).

This might be more precise than Self::unread_notification_counts for encrypted rooms.

pub fn num_unread_mentions(&self) -> u64

Get the number of unread mentions (computed client-side), that is, messages causing a highlight in a room.

This might be more precise than Self::unread_notification_counts for encrypted rooms.

pub fn are_members_synced(&self) -> bool

Check if the room has its members fully synced.

Members might be missing if lazy member loading was enabled for the sync.

Returns true if no members are missing, false otherwise.

pub fn mark_members_missing(&self)

Mark this Room as still missing member information.

pub fn is_state_fully_synced(&self) -> bool

Check if the room states have been synced

States might be missing if we have only seen the room_id of this Room so far, for example as the response for a create_room request without being synced yet.

Returns true if the state is fully synced, false otherwise.

pub fn is_encryption_state_synced(&self) -> bool

Check if the room has its encryption event synced.

The encryption event can be missing when the room hasn’t appeared in sync yet.

Returns true if the encryption state is synced, false otherwise.

pub fn last_prev_batch(&self) -> Option<String>

Get the prev_batch token that was received from the last sync. May be None if the last sync contained the full room history.

pub fn avatar_url(&self) -> Option<OwnedMxcUri>

Get the avatar url of this room.

pub fn canonical_alias(&self) -> Option<OwnedRoomAliasId>

Get the canonical alias of this room.

pub fn alt_aliases(&self) -> Vec<OwnedRoomAliasId>

Get the canonical alias of this room.

pub fn create_content(&self) -> Option<RoomCreateWithCreatorEventContent>

Get the m.room.create content of this room.

This usually isn’t optional but some servers might not send an m.room.create event as the first event for a given room, thus this can be optional.

For room versions earlier than room version 11, if the event is redacted, all fields except creator will be set to their default value.

pub async fn is_direct(&self) -> StoreResult<bool>

Is this room considered a direct message.

pub fn direct_targets(&self) -> HashSet<OwnedUserId>

If this room is a direct message, get the members that we’re sharing the room with.

Note: The member list might have been modified in the meantime and the targets might not even be in the room anymore. This setting should only be considered as guidance. We leave members in this list to allow us to re-find a DM with a user even if they have left, since we may want to re-invite them.

pub fn is_encrypted(&self) -> bool

Is the room encrypted.

pub fn encryption_settings(&self) -> Option<RoomEncryptionEventContent>

Get the m.room.encryption content that enabled end to end encryption in the room.

pub fn guest_access(&self) -> GuestAccess

Get the guest access policy of this room.

pub fn history_visibility(&self) -> HistoryVisibility

Get the history visibility policy of this room.

pub fn is_public(&self) -> bool

Is the room considered to be public.

pub fn join_rule(&self) -> JoinRule

Get the join rule policy of this room.

pub fn max_power_level(&self) -> i64

Get the maximum power level that this room contains.

This is useful if one wishes to normalize the power levels, e.g. from 0-100 where 100 would be the max power level.

pub fn name(&self) -> Option<String>

Get the m.room.name of this room.

The returned string is guaranteed not to be empty.

pub fn is_tombstoned(&self) -> bool

Has the room been tombstoned.

pub fn tombstone(&self) -> Option<RoomTombstoneEventContent>

Get the m.room.tombstone content of this room if there is one.

pub fn topic(&self) -> Option<String>

Get the topic of the room.

pub fn has_active_room_call(&self) -> bool

Is there a non expired membership with application “m.call” and scope “m.room” in this room

pub fn active_room_call_participants(&self) -> Vec<OwnedUserId>

Returns a Vec of userId’s that participate in the room call.

matrix_rtc memberships with application “m.call” and scope “m.room” are considered. A user can occur twice if they join with two devices. convert to a set depending if the different users are required or the amount of sessions.

The vector is ordered by oldest membership user to newest.

pub async fn display_name(&self) -> StoreResult<DisplayName>

Return the cached display name of the room if it was provided via sync, or otherwise calculate it, taking into account its name, aliases and members.

The display name is calculated according to this algorithm.

pub fn latest_event(&self) -> Option<LatestEvent>

Return the last event in this room, if one has been cached during sliding sync.

pub async fn joined_user_ids(&self) -> StoreResult<Vec<OwnedUserId>>

Get the list of users ids that are considered to be joined members of this room.

pub async fn members( &self, memberships: RoomMemberships ) -> StoreResult<Vec<RoomMember>>

Get the RoomMembers of this room that are known to the store, with the given memberships.

pub async fn joined_members(&self) -> StoreResult<Vec<RoomMember>>

👎Deprecated: Use members with RoomMemberships::JOIN instead

Get the list of RoomMembers that are considered to be joined members of this room.

pub async fn active_members(&self) -> StoreResult<Vec<RoomMember>>

👎Deprecated: Use members with RoomMemberships::ACTIVE instead

Get the list of RoomMembers that are considered to be joined or invited members of this room.

pub fn active_members_count(&self) -> u64

Returns the number of members who have joined or been invited to the room.

pub fn invited_members_count(&self) -> u64

Returns the number of members who have been invited to the room.

pub fn joined_members_count(&self) -> u64

Returns the number of members who have joined the room.

pub fn subscribe_info(&self) -> Subscriber<RoomInfo>

Subscribe to the inner RoomInfo.

pub fn clone_info(&self) -> RoomInfo

Clone the inner RoomInfo.

pub fn set_room_info(&self, room_info: RoomInfo)

Update the inner summary with the given RoomInfo, and notify subscribers.

pub async fn get_member( &self, user_id: &UserId ) -> StoreResult<Option<RoomMember>>

Get the RoomMember with the given user_id.

Returns None if the member was never part of this room, otherwise return a RoomMember that can be in a joined, invited, left, banned state.

pub async fn tags(&self) -> StoreResult<Option<Tags>>

Get the Tags for this room.

pub async fn load_user_receipt( &self, receipt_type: ReceiptType, thread: ReceiptThread, user_id: &UserId ) -> StoreResult<Option<(OwnedEventId, Receipt)>>

Get the receipt as an OwnedEventId and Receipt tuple for the given receipt_type, thread and user_id in this room.

pub async fn load_event_receipts( &self, receipt_type: ReceiptType, thread: ReceiptThread, event_id: &EventId ) -> StoreResult<Vec<(OwnedUserId, Receipt)>>

Load from storage the receipts as a list of OwnedUserId and Receipt tuples for the given receipt_type, thread and event_id in this room.