Trait matrix_sdk_base::StateStore

pub trait StateStore: AsyncTraitDeps {

    #[must_use]
    fn save_filter<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        filter_name: &'life1 str, 
        filter_id: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn save_changes<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        changes: &'life1 StateChanges
    ) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_filter<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        filter_name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_sync_token<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_presence_event<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        user_id: &'life1 UserId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<PresenceEvent>>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_state_event<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId, 
        event_type: EventType, 
        state_key: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<AnySyncStateEvent>>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_profile<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId, 
        user_id: &'life2 UserId
    ) -> Pin<Box<dyn Future<Output = Result<Option<MemberEventContent>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_member_event<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId, 
        state_key: &'life2 UserId
    ) -> Pin<Box<dyn Future<Output = Result<Option<MemberEvent>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_user_ids<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<UserId>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_invited_user_ids<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<UserId>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_joined_user_ids<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId
    ) -> Pin<Box<dyn Future<Output = Result<Vec<UserId>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_room_infos<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<RoomInfo>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_stripped_room_infos<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<RoomInfo>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_users_with_display_name<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId, 
        display_name: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<BTreeSet<UserId>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_account_data_event<'life0, 'async_trait>(
        &'life0 self, 
        event_type: EventType
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<AnyBasicEvent>>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_room_account_data_event<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        room_id: &'life1 RoomId, 
        event_type: EventType
    ) -> Pin<Box<dyn Future<Output = Result<Option<Raw<AnyBasicEvent>>, StoreError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

}

An abstract state store trait that can be used to implement different stores for the SDK.

Required methods

#[must_use]
fn save_filter<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
filter_name: &'life1 str,
filter_id: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Save the given filter id under the given name.

Arguments

filter_name - The name that should be used to store the filter id.

filter_id - The filter id that should be stored in the state store.

#[must_use]
fn save_changes<'life0, 'life1, 'async_trait>(
&'life0 self,
changes: &'life1 StateChanges
) -> Pin<Box<dyn Future<Output = Result<(), StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Save the set of state changes in the store.

#[must_use]
fn get_filter<'life0, 'life1, 'async_trait>(
&'life0 self,
filter_name: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Option<String>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get the filter id that was stored under the given filter name.

Arguments

filter_name - The name that was used to store the filter id.

#[must_use]
fn get_sync_token<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<String>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the last stored sync token.

#[must_use]
fn get_presence_event<'life0, 'life1, 'async_trait>(
&'life0 self,
user_id: &'life1 UserId
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<PresenceEvent>>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get the stored presence event for the given user.

Arguments

user_id - The id of the user for which we wish to fetch the presence event for.

#[must_use]
fn get_state_event<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId,
event_type: EventType,
state_key: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<AnySyncStateEvent>>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Get a state event out of the state store.

Arguments

room_id - The id of the room the state event was received for.

event_type - The event type of the state event.

#[must_use]
fn get_profile<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId,
user_id: &'life2 UserId
) -> Pin<Box<dyn Future<Output = Result<Option<MemberEventContent>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Get the current profile for the given user in the given room.

Arguments

room_id - The room id the profile is used in.

user_id - The id of the user the profile belongs to.

#[must_use]
fn get_member_event<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId,
state_key: &'life2 UserId
) -> Pin<Box<dyn Future<Output = Result<Option<MemberEvent>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Get a raw MemberEvent for the given state key in the given room id.

Arguments

room_id - The room id the member event belongs to.

state_key - The user id that the member event defines the state for.

#[must_use]
fn get_user_ids<'life0, 'life1, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId
) -> Pin<Box<dyn Future<Output = Result<Vec<UserId>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get all the user ids of members for a given room.

#[must_use]
fn get_invited_user_ids<'life0, 'life1, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId
) -> Pin<Box<dyn Future<Output = Result<Vec<UserId>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get all the user ids of members that are in the invited state for a given room.

#[must_use]
fn get_joined_user_ids<'life0, 'life1, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId
) -> Pin<Box<dyn Future<Output = Result<Vec<UserId>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get all the user ids of members that are in the joined state for a given room.

#[must_use]
fn get_room_infos<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<RoomInfo>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get all the pure RoomInfos the store knows about.

#[must_use]
fn get_stripped_room_infos<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<RoomInfo>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get all the pure RoomInfos the store knows about.

#[must_use]
fn get_users_with_display_name<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId,
display_name: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<BTreeSet<UserId>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Get all the users that use the given display name in the given room.

Arguments

room_id - The id of the room for which the display name users should be fetched for.

display_name - The display name that the users use.

#[must_use]
fn get_account_data_event<'life0, 'async_trait>(
&'life0 self,
event_type: EventType
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<AnyBasicEvent>>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get an event out of the account data store.

Arguments

event_type - The event type of the account data event.

#[must_use]
fn get_room_account_data_event<'life0, 'life1, 'async_trait>(
&'life0 self,
room_id: &'life1 RoomId,
event_type: EventType
) -> Pin<Box<dyn Future<Output = Result<Option<Raw<AnyBasicEvent>>, StoreError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get an event out of the room account data store.

Arguments

room_id - The id of the room for which the room account data event should be fetched.

event_type - The event type of the room account data event.

Loading content...